separando funcion get status con betsy
