+++ src/view-controller/router.js
@@ -1,5 +1,5 @@
 import { components } from '../view/index.js';
-import { getStatus } from '../firestore.js';
+import { getStatus } from '../firestore-controller/firestore.js';
 
 // controlar el cambio de la vista == view-controler
 export const changeTmp = (hash) => {
